pyproject.toml
[tool.poetry]
name = "overeasy"
version = "0.0.2"
description = ""
authors = ["Ani Rahul <anirahul@redshiftai.app>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.10, <3.13"
torch = "^2.0.0"
numpy = "^1.24.3"
opencv-python-headless = "^4.7.0"
Pillow = "^10.2.0"
requests = "^2.32.0"
matplotlib = "^3.8.1"
anthropic = "^0.25.0"
openai = "^1.12.0"
tabulate = "^0.9.0"
transformers = "^4.29.2"
ultralytics = "^8.2.0"
supervision = "^0.18.0"
pydantic = "^2.0.0"
pydantic_numpy = "^5.0.0"
rf_groundingdino = "0.1.2"
rf_segment_anything = "1.0.0"
tiktoken = "^0.4.0"
google-generativeai = "^0.5.0"
open-clip-torch = "^2.0.0"
tqdm = "^4.65.0"
einops = "^0.6.0"
transformers-stream-generator = "0.0.3"
accelerate = "^0.27.0"
instructor = "^1.3.2"
gradio = "4.25.0"
boto3 = "^1.34.121"

[tool.poetry.scripts]
post_install = "overeasy.models.detection.detic:check_dependencies"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
